requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
matplotlib
numpy